pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tunnelgraf"
version = "0.1.1"
dependencies = [
  "click>=8.0.3",
  "sshtunnel>=0.4.0",
  "pyyaml>=6.0.1",
  "pyinstaller>=6.3.0",
  "python-hosts>=1.0.5",
  "pydantic>=2.5.3",
]
authors = [
  {name = "Dennis Walker", email = "denniswalker@me.com"},
]
maintainers = [
  {name = "Dennis Walker", email = "denniswalker@me.com"}
]
description = "Hierarchical SSH tunnel management made easy. Securely define and share multihop connection profiles that expose a graph of remote endpoints as localhost."
license = {text = "MIT License"}
keywords = ["ssh", "tunnel", "hosts", "connectivity", "secrets", "dns", "remote"]
classifiers = [
  # How mature is this project? Common values are
  #   3 - Alpha
  #   4 - Beta
  #   5 - Production/Stable
  "Development Status :: 3 - Alpha",

  # Indicate who your project is intended for
  "Intended Audience :: System Administrators",
  "Topic :: System :: Networking",

  # Pick your license as you wish (see also "license" above)
  "License :: OSI Approved :: MIT License",

  # Specify the Python versions you support here.
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
]

[project.scripts]
tunnelgraf = "tunnelgraf:main"

[project.urls]
Homepage = "https://github.com/denniswalker/tunnelgraf"
Documentation = "https://github.com/denniswalker/tunnelgraf/blob/main/README.md"
Repository = "https://github.com/denniswalker/tunnelgraf"
Issues = "https://github.com/denniswalker/tunnelgraf/issues"

[tool.hatch.envs.test]
dependencies = [
  "coverage[toml]",
  "pytest",
  "pytest-cov",
  "pytest-mock",
]

[tool.hatch.envs.test.scripts]
run-coverage = "pytest --cov-config=pyproject.toml --cov=pkg --cov=tests"
run = "run-coverage --no-cov"

[tool.hatch.envs.style]
detached = true
dependencies = [
  "flake8",
  "black",
  "isort",
]

[tool.hatch.envs.style.scripts]
check = [
  "flake8 .",
  "black --check --diff .",
  "isort --check-only --diff .",
]
fmt = [
  "isort .",
  "black .",
  "check",
]

[tool.pyright]
include = ["src"]
exclude = ["**/node_modules",
    "**/__pycache__",
]
defineConstant = { DEBUG = true }

reportMissingImports = true
reportMissingTypeStubs = true

executionEnvironments = [
  { root = "src/tunnelgraf" },
  { root = "src/tests" },
  { root = "src" }
]
